backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
sqlalchemy==2.0.38
aiosqlite==0.21.0
asyncpg==0.30.0
pydantic-settings==2.8.1
python-dotenv==1.0.1
email-validator==2.2.0
google-api-python-client==2.165.0
google-auth==2.38.0

